+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Compatibility:
 
 * `GC.stat` can now take an option (#1716).
 * `Kernel#load` with `wrap` has been implemented (#1739).
+* Implemented `Kernel#spawn` with `:chdir` (#1492).
 
 # 19.2.0